# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SH3 Backend API",
        "description": "API de Atendimentos feita por Hugo Galli para SH3",
        "contact": {
            "name": "Hugo Galli",
            "email": "hugo.galli.r@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/analistas/{analistaId}/areas/{areaId}": {
            "post": {
                "tags": [
                    "Analista-Áreas"
                ],
                "summary": "Associar analista a área",
                "description": "Associa um analista a uma área específica. ",
                "operationId": "associateAnalistaArea",
                "parameters": [
                    {
                        "name": "analistaId",
                        "in": "path",
                        "description": "ID do analista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "areaId",
                        "in": "path",
                        "description": "ID da área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Analista associated with area successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Analista ou área não encontrados"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        },
        "/api/analista-areas/dissociate/{analistaId}/{areaId}": {
            "post": {
                "tags": [
                    "Analista-Áreas"
                ],
                "summary": "Desassociar analista de área",
                "description": "Desassocia um analista de uma área específica.",
                "operationId": "dissociateAnalistaArea",
                "parameters": [
                    {
                        "name": "analistaId",
                        "in": "path",
                        "description": "ID do analista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "areaId",
                        "in": "path",
                        "description": "ID da área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Analista dissociated from area successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Analista ou área não encontrados"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "jwt_token": {
                "type": "apiKey",
                "scheme": "bearer",
                "description": "Insira valores no formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}